Replace ZKSProvider with ZKMiddleware
